read url to bytes buffer instead of a string
RSS feed can be encoded in different encodings than UTF-8 and the
quick_xml library is capable of handling this feature. The previous
implementation prevented this from working by loading the
response into UTF-8 encoded string, which failed. By using a byte array
instead of the string we can use the feature.